+++ src/components/Card.js
@@ -1,4 +1,9 @@
-import { CardType, smallScale, cardHeight, cardWidth } from './utils/constants';
+import {
+  CardType,
+  cardSmallScale,
+  cardHeight,
+  cardWidth,
+} from './utils/constants';
 import { cardFront, cardBack, cardPlaceholder } from './utils/assetPaths';
 
 const Card = ({
@@ -7,7 +12,7 @@ const Card = ({
   cardIndex,
   playerId = '0',
   handleCardClick = () => {},
-  scale = smallScale,
+  scale = cardSmallScale,
 }) => {
   const frontImg = cardFront(cardId);
   const backImg = cardBack(playerId);