+++ src/components/EffectStack.js
@@ -1,7 +1,27 @@
-import { sortEffects } from './utils/utils';
 import { icon } from './utils/assetPaths';
 
 const EffectStack = ({ opponentEffects, playerEffects, showEffectStack }) => {
+  const sortEffects = (effects) => {
+    // Sort by group first (buff then debuff), then name (group same non-unique effects together)
+    return effects.sort((a, b) => {
+      if (a.group === 'buff' && b.group === 'debuff') {
+        return -1;
+      } else if (a.group === 'debuff' && b.group === 'buff') {
+        return 1;
+      } else if (a.group === b.group) {
+        if (a.text < b.text) {
+          return -1;
+        } else if (a.text > b.text) {
+          return 1;
+        } else {
+          return 0;
+        }
+      } else {
+        return 0; // Keep the original order if groups are the same but texts are different
+      }
+    });
+  };
+
   return (
     showEffectStack && (
       <div className='d-flex flex-column h-100 justify-content-between'>