+++ src/components/PlayerStatsPanel.js
@@ -1,13 +1,22 @@
 import { useState } from 'react';
-import { smallScale, avatarHeight, avatarWidth } from './utils/constants';
+import {
+  avatarHeight,
+  avatarWidth,
+  avatarMediumScale,
+} from './utils/constants';
 import { getAvatarForLevel, icon } from './utils/assetPaths';
 import { EffectType } from '../data/cardEffects';
 
-const PlayerStatsPanel = ({ player, level, scale = smallScale }) => {
+const PlayerStatsPanel = ({
+  player,
+  level,
+  showCardAnimation,
+  cardAnimationData,
+}) => {
   const [isStatsIconsHovered, setIsStatsIconsHovered] = useState(false);
 
-  const height = avatarHeight * scale;
-  const width = avatarWidth * scale;
+  const height = avatarHeight * avatarMediumScale;
+  const width = avatarWidth * avatarMediumScale;
 
   const formatValueDisplay = (currentValue, maxValue) => {
     if (currentValue >= maxValue) {
@@ -20,11 +29,12 @@ const PlayerStatsPanel = ({ player, level, scale = smallScale }) => {
   return (
     <div className='d-flex align-items-center'>
       <div
-        className='me-2 position-relative'
+        className='me-3 position-relative'
         data-bs-toggle='tooltip'
         data-bs-placement={player.id === '0' ? 'top' : 'bottom'}
         data-bs-title={player.id === '0' ? 'Player' : 'Opponent'}
       >
+        {/* Player avatar */}
         <img
           src={getAvatarForLevel(player.id, level)}
           alt='avatar'
@@ -32,26 +42,37 @@ const PlayerStatsPanel = ({ player, level, scale = smallScale }) => {
           width={width}
         />
 
+        {/* Image overlay for level effect */}
         {player.effects.some((e) => e.type === EffectType.freeze) && (
-          <img src={icon.ice} alt='ice' className='ice-overlay' />
+          <img src={icon.ice} alt='ice' className='overlay-ice' />
+        )}
+
+        {/* Animation overlay for card effect */}
+        {showCardAnimation && (
+          <img
+            src={cardAnimationData?.path}
+            alt={cardAnimationData?.type}
+            className={`vfx-${cardAnimationData?.type}`}
+          />
         )}
       </div>
 
+      {/* Player stats (hp, atk, def) */}
       <div
         className='d-flex flex-column'
         onMouseEnter={() => setIsStatsIconsHovered(true)}
         onMouseLeave={() => setIsStatsIconsHovered(false)}
       >
         <div className='d-flex align-items-center mb-2'>
-          <img src={icon.hp} className='me-2' alt='hp' />
-          <span className='fw-semibold'>
+          <img src={icon.hp} className='me-2 stats-icon' alt='hp' />
+          <span className='fw-semibold stats-text'>
             {player.hp} {isStatsIconsHovered && `/ ${player.maxHp}`}
           </span>
         </div>
 
         <div className='d-flex align-items-center mb-2'>
-          <img src={icon.atk} className='me-2' alt='atk' />
-          <span className='fw-semibold'>
+          <img src={icon.atk} className='me-2 stats-icon' alt='atk' />
+          <span className='fw-semibold stats-text'>
             {isStatsIconsHovered
               ? formatValueDisplay(player.atk, player.baseAtk)
               : player.atk}
@@ -59,8 +80,8 @@ const PlayerStatsPanel = ({ player, level, scale = smallScale }) => {
         </div>
 
         <div className='d-flex align-items-center'>
-          <img src={icon.def} className='me-2' alt='def' />
-          <span className='fw-semibold'>
+          <img src={icon.def} className='me-2 stats-icon' alt='def' />
+          <span className='fw-semibold stats-text'>
             {isStatsIconsHovered
               ? formatValueDisplay(player.def, player.baseDef)
               : player.def}