+++ src/components/utils/assetPaths.js
@@ -1,7 +1,10 @@
+import { AnimationTarget } from './constants';
+
 const imgPrefix = 'images';
 const avatarPrefix = `${imgPrefix}/avatars`;
 const cardPrefix = `${imgPrefix}/cards`;
 const locationPrefix = `${imgPrefix}/locations`;
+const animationPrefix = 'animations';
 const audioPrefix = 'audio';
 const musicPrefix = 'music';
 
@@ -15,6 +18,7 @@ const avatarPaths = {
   4: `${avatarPrefix}/traveling-merchant.svg`,
   5: `${avatarPrefix}/forest-ranger.svg`,
   6: `${avatarPrefix}/murloc-oracle.svg`,
+  7: `${avatarPrefix}/crimson-witch.svg`,
 };
 const getAvatarForLevel = (playerId, level = '1') => {
   return playerId === '0' ? avatarPaths[0] : avatarPaths[level];
@@ -27,6 +31,7 @@ const locationPaths = {
   4: `${locationPrefix}/desert.svg`,
   5: `${locationPrefix}/forest.svg`,
   6: `${locationPrefix}/ocean.svg`,
+  7: `${locationPrefix}/marsh.svg`,
 };
 const getLocationForLevel = (level = '1') => locationPaths[level];
 
@@ -52,7 +57,7 @@ const cardPlaceholder = (playerId) =>
   `${cardPrefix}/placeholder/${playerId}.svg`;
 const cardPile = `${cardPrefix}/pile.svg`;
 
-const cardCount = 34; // Increment this when new cards are added
+const cardCount = 36; // Increment this when new cards are added
 const cardFronts = Array.from(
   { length: cardCount },
   (_, cardId) => `${cardPrefix}/front/${cardId}.svg`
@@ -64,8 +69,251 @@ const nextLevelCards = {
   3: [cardFront('24'), cardFront('25')],
   4: [cardFront('30'), cardFront('31')],
   5: [cardFront('32'), cardFront('33')],
+  6: [cardFront('34'), cardFront('35')],
 };
 const getNextCardsForLevel = (level = '1') => nextLevelCards[level];
+
+// ANIMATION FILES
+
+const getAnimationPath = (type) => `${animationPrefix}/${type}.gif`;
+const animationData = {
+  fireball: {
+    type: 'fireball',
+    path: getAnimationPath('fireball'),
+    timeout: 500,
+  },
+  freeze: {
+    type: 'freeze',
+    path: getAnimationPath('freeze'),
+    timeout: 1000,
+  },
+  thunder: {
+    type: 'thunder',
+    path: getAnimationPath('thunder'),
+    timeout: 1000,
+  },
+  heal: {
+    type: 'heal',
+    path: getAnimationPath('heal'),
+    timeout: 500,
+  },
+  sword: {
+    type: 'sword',
+    path: getAnimationPath('sword'),
+    timeout: 1000,
+  },
+  shield: {
+    type: 'shield',
+    path: getAnimationPath('shield'),
+    timeout: 1200,
+  },
+  'ghost-bounce': {
+    type: 'ghost-bounce',
+    path: getAnimationPath('ghost-bounce'),
+    timeout: 1200,
+  },
+  skull: {
+    type: 'skull',
+    path: getAnimationPath('skull'),
+    timeout: 1800,
+  },
+  hammer: {
+    type: 'hammer',
+    path: getAnimationPath('hammer'),
+    timeout: 1000,
+  },
+  'blood-strike': {
+    type: 'blood-strike',
+    path: getAnimationPath('blood-strike'),
+    timeout: 1000,
+  },
+  'blood-burst': {
+    type: 'blood-burst',
+    path: getAnimationPath('blood-burst'),
+    timeout: 1000,
+  },
+  'star-bounce': {
+    type: 'star-bounce',
+    path: getAnimationPath('star-bounce'),
+    timeout: 1200,
+  },
+  'star-strike': {
+    type: 'star-strike',
+    path: getAnimationPath('star-strike'),
+    timeout: 1000,
+  },
+  'star-impact': {
+    type: 'star-impact',
+    path: getAnimationPath('star-impact'),
+    timeout: 600,
+  },
+  'slime-swirl': {
+    type: 'slime-swirl',
+    path: getAnimationPath('slime-swirl'),
+    timeout: 1000,
+  },
+  'slime-strike': {
+    type: 'slime-strike',
+    path: getAnimationPath('slime-strike'),
+    timeout: 600,
+  },
+  'water-explode': {
+    type: 'water-explode',
+    path: getAnimationPath('water-explode'),
+    timeout: 800,
+  },
+  'blood-cross-strike': {
+    type: 'blood-cross-strike',
+    path: getAnimationPath('blood-cross-strike'),
+    timeout: 800,
+  },
+  'slime-splash': {
+    type: 'slime-splash',
+    path: getAnimationPath('slime-splash'),
+    timeout: 500,
+  },
+};
+const cardAnimation = {
+  0: {
+    type: 'fireball',
+    target: AnimationTarget.enemy,
+  },
+  1: {
+    type: 'fireball',
+    target: AnimationTarget.enemy,
+  },
+  2: {
+    type: 'fireball',
+    target: AnimationTarget.enemy,
+  },
+  3: {
+    type: 'freeze',
+    target: AnimationTarget.enemy,
+  },
+  4: {
+    type: 'freeze',
+    target: AnimationTarget.enemy,
+  },
+  5: {
+    type: 'freeze',
+    target: AnimationTarget.enemy,
+  },
+  6: {
+    type: 'thunder',
+    target: AnimationTarget.enemy,
+  },
+  7: {
+    type: 'thunder',
+    target: AnimationTarget.enemy,
+  },
+  8: {
+    type: 'thunder',
+    target: AnimationTarget.enemy,
+  },
+  9: {
+    type: 'heal',
+    target: AnimationTarget.self,
+  },
+  10: {
+    type: 'heal',
+    target: AnimationTarget.self,
+  },
+  11: {
+    type: 'heal',
+    target: AnimationTarget.self,
+  },
+  12: {
+    type: 'sword',
+    target: AnimationTarget.self,
+  },
+  13: {
+    type: 'shield',
+    target: AnimationTarget.self,
+  },
+  14: {
+    type: 'ghost-bounce',
+    target: AnimationTarget.enemy,
+  },
+  15: {
+    type: 'skull',
+    target: AnimationTarget.enemy,
+  },
+  16: {
+    type: 'hammer',
+    target: AnimationTarget.self,
+  },
+  17: {
+    type: 'blood-strike',
+    target: AnimationTarget.enemy,
+  },
+  18: {
+    type: 'sword',
+    target: AnimationTarget.self,
+  },
+  19: {
+    type: 'shield',
+    target: AnimationTarget.self,
+  },
+  20: {
+    type: 'fireball',
+    target: AnimationTarget.enemy,
+  },
+  21: {
+    type: 'blood-burst',
+    target: AnimationTarget.self,
+  },
+  22: {
+    type: 'freeze',
+    target: AnimationTarget.enemy,
+  },
+  23: {
+    type: 'heal',
+    target: AnimationTarget.self,
+  },
+  24: null, 
+  25: null,
+  26: {
+    type: 'star-bounce',
+    target: AnimationTarget.enemy,
+  },
+  27: {
+    type: 'star-strike',
+    target: AnimationTarget.self,
+  },
+  28: {
+    type: 'star-bounce',
+    target: AnimationTarget.self,
+  },
+  29: {
+    type: 'star-impact',
+    target: AnimationTarget.enemy,
+  },
+  30: {
+    type: 'slime-swirl',
+    target: AnimationTarget.both,
+  },
+  31: {
+    type: 'slime-strike',
+    target: AnimationTarget.enemy,
+  },
+  32: null, 
+  33: {
+    type: 'water-explode',
+    target: AnimationTarget.enemy,
+  },
+  34: {
+    type: 'blood-cross-strike',
+    target: AnimationTarget.self,
+  },
+  35: {
+    type: 'slime-splash',
+    target: AnimationTarget.enemy,
+  },
+};
+const getAnimationDataForCard = (cardId) =>
+  animationData[cardAnimation[cardId]?.type];
+// Same animation can be reused to different targets, hence separating it from animationData object
+const getAnimationTargetForCard = (cardId) => cardAnimation[cardId]?.target;
 
 // AUDIO FILES
 
@@ -104,6 +352,8 @@ const cardAudioType = {
   31: 'arrow',
   32: 'waves',
   33: 'waves',
+  34: 'roar',
+  35: 'potion',
 };
 const cardAudio = (cardId) => `${audioPrefix}/${cardAudioType[cardId]}.mp3`;
 const click = `${audioPrefix}/click.mp3`;
@@ -112,6 +362,7 @@ const defeat = `${audioPrefix}/defeat.mp3`;
 const miss = `${audioPrefix}/wind.mp3`; // reuse the wind sound effect
 const defrost = `${audioPrefix}/freeze.mp3`; // reuse the freeze sound effect
 const cleanse = `${audioPrefix}/waves.mp3`; // reuse the waves sound effect
+const potion = `${audioPrefix}/potion.mp3`; // reuse the potion sound effect
 
 // MUSIC FILES
 
@@ -122,6 +373,7 @@ const musicPaths = {
   4: `${musicPrefix}/lost-desert.mp3`,
   5: `${musicPrefix}/misty-woods.mp3`,
   6: `${musicPrefix}/coral-bay.mp3`,
+  7: `${musicPrefix}/shadow-swamp.mp3`,
 };
 const getMusicForLevel = (level = '1') => musicPaths[level];
 
@@ -138,6 +390,7 @@ const IMAGES = [
   `${cardPrefix}/placeholder/1.svg`,
   cardPile,
 ];
+const ANIMATION = Object.values(animationData).map((data) => data.path);
 const AUDIO = [
   click,
   victory,
@@ -153,6 +406,8 @@ const AUDIO = [
   `${audioPrefix}/wind.mp3`,
   `${audioPrefix}/arrow.mp3`,
   `${audioPrefix}/waves.mp3`,
+  `${audioPrefix}/roar.mp3`,
+  `${audioPrefix}/potion.mp3`,
 ];
 const MUSIC = [...Object.values(musicPaths)];
 
@@ -168,6 +423,10 @@ export {
   cardPile,
   cardFronts,
 
+  // ANIMATION FILES
+  getAnimationDataForCard,
+  getAnimationTargetForCard,
+
   // AUDIO FILES
   cardAudio,
   click,
@@ -176,12 +435,14 @@ export {
   miss,
   defrost,
   cleanse,
+  potion,
 
   // MUSIC FILES
   getMusicForLevel,
 
   // ALL FILES FOR PRELOAD
   IMAGES,
+  ANIMATION,
   AUDIO,
   MUSIC,
 };