+++ src/game/effectUtils.js
@@ -95,6 +95,10 @@ export const undoEffect = (G, target, { type, value = 0 }) => {
       break;
     case EffectType.aura:
       break;
+    case EffectType.counterAttack:
+      break;
+    case EffectType.poison:
+      break;
     default:
       console.error(`Invalid effect type: ${type}`);
       return;